+++ system/src/Grav/Installer/InstallException.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * @package    Grav\Installer
+ *
+ * @copyright  Copyright (C) 2015 - 2020 Trilby Media, LLC. All rights reserved.
+ * @license    MIT License; see LICENSE file for details.
+ */
+
+namespace Grav\Installer;
+
+use Throwable;
+
+/**
+ * Class InstallException
+ * @package Grav\Installer
+ */
+class InstallException extends \RuntimeException
+{
+    /**
+     * InstallException constructor.
+     * @param string $message
+     * @param Throwable $previous
+     */
+    public function __construct(string $message, Throwable $previous)
+    {
+        parent::__construct($message, $previous->getCode(), $previous);
+    }
+}